requirements.txt
langchain>=0.2.7
langchain-community>=0.2.7
langchain-text-splitters>=0.2.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=4.2.0
requests>=2.31.0
openai>=1.37.0
flask>=2.3.2
flask-cors>=3.0.10
flask-wtf>=1.0.1
flask-login>=0.6.2
flask-mail>=0.9.1
flask-sqlalchemy>=3.0.3
flask-migrate>=4.0.4
flask-bcrypt>=1.0.1
flask-jwt-extended>=4.5.0
flask-caching>=2.0.0
flask-restful>=0.3.9
langchain-openai>=1.0.2
python-dotenv>=1.0.0